refactor: 🔨 Move cli logic to commands because it makes more senses
