--- src/cli.rs
@@ -1,31 +1,0 @@
-// SPDX-FileCopyrightText: 2023 The MalwareTracesGenerator development team
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-use crate::commands::{generate::Generate, traces::Traces, Runnable};
-use clap::{Parser, Subcommand};
-use std::error::Error;
-
-#[derive(Parser)]
-#[clap(author, version)]
-#[clap(arg_required_else_help = true)]
-pub struct Arguments {
-    #[clap(subcommand)]
-    pub command: Command,
-}
-
-#[derive(Subcommand)]
-pub enum Command {
-    Traces(Traces),
-    Generate(Generate),
-}
-
-impl Runnable for Arguments {
-    fn run(&self) -> Result<(), Box<dyn Error>> {
-        match &self.command {
-            Command::Traces(traces) => traces as &dyn Runnable,
-            Command::Generate(generate) => generate,
-        }
-        .run()
-    }
-}
