+++ src/commands.rs
@@ -2,11 +2,37 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use std::error::Error;
-
 pub mod generate;
 pub mod traces;
 
+use crate::commands::{generate::Generate, traces::Traces};
+use clap::{Parser, Subcommand};
+use std::error::Error;
+
+#[derive(Parser)]
+#[clap(author, version)]
+#[clap(arg_required_else_help = true)]
+pub struct Arguments {
+    #[clap(subcommand)]
+    pub command: Command,
+}
+
+#[derive(Subcommand)]
+pub enum Command {
+    Traces(Traces),
+    Generate(Generate),
+}
+
 pub trait Runnable {
     fn run(&self) -> Result<(), Box<dyn Error>>;
+}
+
+impl Runnable for Arguments {
+    fn run(&self) -> Result<(), Box<dyn Error>> {
+        match &self.command {
+            Command::Traces(traces) => traces as &dyn Runnable,
+            Command::Generate(generate) => generate,
+        }
+        .run()
+    }
 }
