+++ src/main.rs
@@ -2,13 +2,11 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-mod cli;
 mod commands;
 mod windows;
 
 use clap::Parser;
-use cli::Arguments;
-use commands::Runnable;
+use commands::{Arguments, Runnable};
 use std::error::Error;
 
 fn main() -> Result<(), Box<dyn Error>> {
